Move controller tests into tests/controllers
