+++ tests/controllers/missing-delivery-reports.js
@@ -1,6 +1,6 @@
-var controller = require('../controllers/missed-appointments'),
-    db = require('../db'),
-    config = require('../config'),
+var controller = require('../../controllers/missing-delivery-reports'),
+    db = require('../../db'),
+    config = require('../../config'),
     moment = require('moment'),
     sinon = require('sinon');
 
@@ -22,6 +22,7 @@ exports.tearDown = function (callback) {
   }
   callback();
 };
+
 
 exports['get returns errors'] = function(test) {
   test.expect(2);
@@ -83,88 +84,8 @@ exports['get returns zero if all registrations have delivered'] = function(test)
   });
 };
 
-exports['get returns zero if all registrations have visits'] = function(test) {
-  test.expect(2);
-  var fti = sinon.stub(db, 'fti');
-  fti.onFirstCall().callsArgWith(2, null, {
-    rows: [
-      { 
-        doc: { 
-          patient_id: 1,
-          scheduled_tasks: [ {
-            group: 1,
-            due: moment().subtract(20, 'days').toISOString()
-          } ]
-        } 
-      },
-      { 
-        doc: { 
-          patient_id: 2,
-          scheduled_tasks: [ {
-            group: 1,
-            due: moment().subtract(20, 'days').toISOString()
-          } ]
-        } 
-      }
-    ]
-  });
-  fti.onSecondCall().callsArgWith(2, null, {
-    rows: [
-      { doc: { patient_id: 1 } }
-    ]
-  });
-  fti.onThirdCall().callsArgWith(2, null, {
-    rows: [
-      { doc: { patient_id: 2 } }
-    ]
-  });
-  controller.get({}, function(err, results) {
-    test.equals(results.length, 0);
-    test.equals(fti.callCount, 3);
-    test.done();
-  });
-};
-
-exports['get ignores registrations with no missed appointments'] = function(test) {
-  test.expect(2);
-  var fti = sinon.stub(db, 'fti');
-  fti.onFirstCall().callsArgWith(2, null, {
-    rows: [
-      { 
-        doc: { 
-          patient_id: 1,
-          scheduled_tasks: []
-        } 
-      },
-      { 
-        doc: { 
-          patient_id: 2,
-          scheduled_tasks: [ {
-            group: 1,
-            due: moment().subtract(24, 'days').toISOString()
-          } ]
-        } 
-      },
-      { 
-        doc: { 
-          patient_id: 3,
-          scheduled_tasks: [ {
-            group: 1,
-            due: moment().subtract(13, 'days').toISOString()
-          } ]
-        } 
-      }
-    ]
-  });
-  controller.get({}, function(err, results) {
-    test.equals(results.length, 0);
-    test.equals(fti.callCount, 1);
-    test.done();
-  });
-};
-
-exports['get returns all registrations with missed appointments'] = function(test) {
-  test.expect(18);
+exports['get returns all registrations with missed delivery reports'] = function(test) {
+  test.expect(14);
   var fti = sinon.stub(db, 'fti');
   var today = moment();
   fti.onFirstCall().callsArgWith(2, null, {
@@ -174,7 +95,7 @@ exports['get returns all registrations with missed appointments'] = function(tes
           patient_id: 1,
           patient_name: 'sarah',
           form: 'R',
-          reported_date: today.clone().subtract(10, 'weeks').toISOString(),
+          reported_date: today.clone().subtract(38, 'weeks').toISOString(),
           related_entities: { clinic: { id: 'x' } },
           scheduled_tasks: [ {
             group: 1,
@@ -194,30 +115,31 @@ exports['get returns all registrations with missed appointments'] = function(tes
             due: moment().subtract(20, 'days').toISOString()
           } ]
         } 
+      },
+      { 
+        doc: { 
+          patient_id: 3,
+          patient_name: 'sharon',
+          form: 'P',
+          lmp_date: today.clone().subtract(42, 'weeks').toISOString(),
+          related_entities: { clinic: { id: 'y' } },
+          scheduled_tasks: [ {
+            group: 1,
+            due: moment().subtract(20, 'days').toISOString()
+          } ]
+        } 
       }
     ]
   });
   fti.onSecondCall().callsArgWith(2, null, {
     rows: [
-      { doc: { patient_id: 4 } }
+      { doc: { patient_id: 2 } }
     ]
   });
   fti.onThirdCall().callsArgWith(2, null, {
     rows: [
-      { doc: { patient_id: 5 } }
-    ]
-  });
-  fti.onCall(3).callsArgWith(2, null, {
-    rows: [
-      { doc: { patient_id: 1 } },
-      { doc: { patient_id: 2 } }
-    ]
-  });
-  fti.onCall(4).callsArgWith(2, null, {
-    rows: [
-      { doc: { patient_id: 1 } },
-      { doc: { patient_id: 1 } },
-      { doc: { patient_id: 3 } }
+      { doc: { patient_id: 2 } },
+      { doc: { patient_id: 1 } }
     ]
   });
   controller.get({}, function(err, results) {
@@ -226,22 +148,18 @@ exports['get returns all registrations with missed appointments'] = function(tes
     test.equals(results[0].patient_id, 1);
     test.equals(results[0].patient_name, 'sarah');
     test.equals(results[0].clinic.id, 'x');
-    test.equals(results[0].weeks.number, 10);
-    test.equals(results[0].weeks.approximate, true);
-    test.equals(results[0].date.toISOString(), today.clone().subtract(20, 'days').toISOString());
-    test.equals(results[0].visits, 1);
+    test.equals(results[0].edd.date.toISOString(), today.clone().add(2, 'weeks').toISOString());
+    test.equals(results[0].edd.approximate, true);
     test.equals(results[0].high_risk, true);
 
-    test.equals(results[1].patient_id, 2);
-    test.equals(results[1].patient_name, 'sally');
+    test.equals(results[1].patient_id, 3);
+    test.equals(results[1].patient_name, 'sharon');
     test.equals(results[1].clinic.id, 'y');
-    test.equals(results[1].weeks.number, 12);
-    test.equals(results[1].weeks.approximate, undefined);
-    test.equals(results[1].date.toISOString(), today.clone().subtract(20, 'days').toISOString());
-    test.equals(results[1].visits, 1);
+    test.equals(results[1].edd.date.toISOString(), today.toISOString());
+    test.equals(results[1].edd.approximate, undefined);
     test.equals(results[1].high_risk, undefined);
 
-    test.equals(fti.callCount, 5);
+    test.equals(fti.callCount, 3);
     test.done();
   });
 };
