+++ tests/controllers/total-births.js
@@ -1,6 +1,6 @@
-var controller = require('../controllers/total-births'),
-    db = require('../db'),
-    config = require('../config'),
+var controller = require('../../controllers/total-births'),
+    db = require('../../db'),
+    config = require('../../config'),
     sinon = require('sinon');
 
 exports.setUp = function(callback) {
