+++ tests/controllers/upcoming-appointments.js
@@ -1,6 +1,6 @@
-var controller = require('../controllers/upcoming-appointments'),
-    db = require('../db'),
-    config = require('../config'),
+var controller = require('../../controllers/upcoming-appointments'),
+    db = require('../../db'),
+    config = require('../../config'),
     moment = require('moment'),
     sinon = require('sinon');
