+++ tests/controllers/visits-complete.js
@@ -1,6 +1,6 @@
-var controller = require('../controllers/visits-completed'),
-    db = require('../db'),
-    config = require('../config'),
+var controller = require('../../controllers/visits-completed'),
+    db = require('../../db'),
+    config = require('../../config'),
     sinon = require('sinon');
 
 exports.setUp = function(callback) {
